feat(cpu): widen IP from u16 to u32 — unblocks kernels at 32-bit addresses
Modern Linux maps its kernel at addresses well above 0xFFFF
(typically 0xC0000000+ for 32-bit, or higher for 64-bit). Until
now `Cpu::ip` was a 16-bit register, so any jump beyond the first
64 KiB wrapped silently. This tick widens `ip` to u32 and threads
the change through every reader/writer.

Changes:

  * `Cpu::ip: u32` — single source of truth.
  * `linear_seg(seg, off)` now takes `off: u32`. All callers that
    pass 16-bit offsets (ModR/M effective addresses, SP-relative,
    string-op SI/DI) cast `as u32` at the call site.
  * Jcc / JMP rel / CALL rel: `rel as i32 as u32` so the
    sign-extended offset wraps around the full 32-bit IP space.
  * Near JMP / RET / RETF / IRET / CALL ptr16:16 / JMP ptr16:16:
    new_ip from a u16 pop or fetch becomes `new_ip as u32`. The
    return-IP pushes truncate IP to `as u16` (matching 16-bit
    gate convention; a 32-bit gate path will widen the push).
  * CpuError variants: ip field widened to u32, display now
    formats as 8 hex digits.
  * Snapshot still emits 2 bytes for IP (v1-v5 compatibility);
    high 16 bits truncate on save and zero-extend on load. A
    future v6 widens this.
  * VM's bios_int13 and bios_int15 cast their u16 BX/DI to u32
    when forming ES:OFFSET linear addresses.

New regression test `ip_register_is_32_bit_and_can_hold_addresses_-
above_64kib` puts a 3-byte MOV+HLT stub at linear 0x12_3450,
seeds IP=0x0012_3450 with CS base 0, and asserts the fetch loop
actually executes from that address — the basic prerequisite for
any modern-kernel jump.

All 254 prior tests still pass with no behavioural changes; the
only architectural change is that addresses above 0xFFFF are now
reachable when CS base + IP demand it.

255 tests green (cpu 155 +1). fmt + clippy --workspace --all-targets
-D warnings clean.

+++ crates/vm/src/lib.rs
@@ -139,7 +139,7 @@ fn bios_int13(cpu: &mut Cpu, mem: &mut Memory, io: &mut IoBus) -> bool {
             let lba = (cyl * 2 + head) * 18 + sector;
             // Destination linear = ES.base + BX. read_r16(3) returns BX.
             let bx = cpu.read_r16(3);
-            let dest_linear = cpu.linear_seg(wwwvm_cpu::sreg::ES, bx);
+            let dest_linear = cpu.linear_seg(wwwvm_cpu::sreg::ES, bx as u32);
 
             let mut buf = vec![0u8; count * wwwvm_devices::DISK_SECTOR_SIZE];
             io.disk.read_sectors(lba, count as u8, &mut buf);
@@ -211,7 +211,7 @@ fn bios_int15(cpu: &mut Cpu, mem: &mut Memory) -> bool {
 
     // Linear destination = ES:DI.
     let di = cpu.regs[wwwvm_cpu::r16::DI];
-    let dest = cpu.linear_seg(wwwvm_cpu::sreg::ES, di);
+    let dest = cpu.linear_seg(wwwvm_cpu::sreg::ES, di as u32);
 
     let base: u64 = 0;
     let length = mem.size() as u64;
@@ -451,7 +451,10 @@ impl Vm {
         for s in &self.cpu.sregs {
             buf.extend_from_slice(&s.to_le_bytes());
         }
-        buf.extend_from_slice(&self.cpu.ip.to_le_bytes());
+        // Snapshot still emits 16-bit IP for v1-v5 compatibility.
+        // The Cpu field is now u32; high bits get truncated on save
+        // (a future v6 layout widens this).
+        buf.extend_from_slice(&(self.cpu.ip as u16).to_le_bytes());
         buf.extend_from_slice(&self.cpu.flags.to_le_bytes());
         buf.push(self.cpu.halted as u8);
         buf.push(match self.cpu.seg_override() {
@@ -624,7 +627,7 @@ impl Vm {
         self.cpu.regs = regs;
         self.cpu.regs_high = regs_high;
         self.cpu.sregs = sregs;
-        self.cpu.ip = ip;
+        self.cpu.ip = ip as u32;
         self.cpu.flags = flags;
         self.cpu.halted = halted;
         self.cpu.set_seg_override(seg_override);
